+++ packages/cadl-ranch-specs/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @azure-tools/cadl-ranch-specs
 
+## 0.34.9
+
+### Patch Changes
+
+- Updated dependencies [1c7eaef]
+  - @azure-tools/cadl-ranch-api@0.4.4
+
+## 0.34.8
+
+### Patch Changes
+
+- cee571f: Add spread-Alias-With-Optional-Properties test cases for type/parameters/spread
+  Add spread-Alias-With-Optional-Collections test cases for type/parameters/spread
+  Add spread-Alias-With-Model test cases for type/parameters/spread
+- 37dfce1: Add xml basic tests
+
 ## 0.34.7
 
 ### Patch Changes